+++ console/src/main/java/com/ctrip/framework/drc/console/service/impl/TransferServiceImpl.java
@@ -1,4 +1,5 @@
 package com.ctrip.framework.drc.console.service.impl;
+import com.ctrip.framework.drc.console.config.DefaultConsoleConfig;
 import com.ctrip.framework.drc.console.dao.entity.*;
 import com.ctrip.framework.drc.console.enums.BooleanEnum;
 import com.ctrip.framework.drc.console.enums.EstablishStatusEnum;
@@ -37,6 +38,8 @@ public class TransferServiceImpl implements TransferService {
     private MetaInfoServiceImpl metaInfoService;
     @Autowired
     private DrcDoubleWriteService drcDoubleWriteService;
+    @Autowired
+    private DefaultConsoleConfig defaultConsoleConfig;
 
     private DalUtils dalUtils = DalUtils.getInstance();
 
@@ -248,7 +251,11 @@ public class TransferServiceImpl implements TransferService {
                     .collect(Collectors.toList());
             srcMha.addAll(destMha);
             doRemove(mhaGroupTbl, srcMha);
-            drcDoubleWriteService.deleteMhaReplicationConfig(srcMha.get(0).getId(), srcMha.get(1).getId());
+            if (defaultConsoleConfig.getDrcDoubleWriteSwitch().equals(DefaultConsoleConfig.SWITCH_ON)) {
+                logger.info("drcDoubleWrite deleteMhaReplicationConfig");
+                drcDoubleWriteService.deleteMhaReplicationConfig(srcMha.get(0).getId(), srcMha.get(1).getId());
+            }
+
         } else {
             throw new Exception("no such mha: " + mhaName);
         }
